backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg[binary]==3.2.3
pgvector==0.3.6
httpx==0.27.2
redis==5.2.0
celery==5.4.0
langgraph==0.2.39
opensearch-py==2.7.1
neo4j==5.25.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
tenacity==9.0.0
orjson==3.10.10
structlog==24.4.0
prometheus-client==0.21.0
opentelemetry-api==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-exporter-otlp==1.28.1
opentelemetry-instrumentation-fastapi==0.49b1
reportlab==4.2.5
pypdf==5.1.0
pytest==8.3.3
pytest-asyncio==0.24.0
respx==0.21.1
psycopg2-binary==2.9.9
cachetools==5.5.0
beautifulsoup4==4.12.3
lxml==5.3.0